requirements.txt

python-dotenv
instagrapi
opencv-python
openai